Source code file
Theme: Pharma Supply Chain System using Smart Contracts

Problem statement: The pharmaceutical industry faces significant challenges related to the security and integrity of the supply chain, including issues such as counterfeiting, diversion, and theft. These challenges can compromise patient safety, reduce the effectiveness of drugs, and cause financial losses for companies. In addition, traditional paper-based supply chain management systems are outdated and inefficient, increasing the likelihood of errors and delays in the supply chain. There is a need for a more secure and transparent system that can enable efficient tracking and tracing of drugs from the point of manufacture to the point of consumption

Goal: The system helps to create a transparent and immutable record of every transaction in the supply chain.

It helps to prevent fraud, counterfeiting, and other illegal activities.

It can reduce the time and resources needed to manage the supply chain.

It helps to improve efficiency and reduce costs.

It assists to prevent hacking, data breaches, and other security threats.